fix(app): address remix breaking changes
This patch replaces a bunch of type names and other minor type fixes to
ensure that the front-end is compatible with Remix v2.

The patch includes the output of running:

```
$ ruplacer "V2_" "" --go
$ ruplacer "RouteMatch" "UIMatch" --go
$ ruplacer "LoaderArgs" "DataFunctionArgs" --go
$ ruplacer "ActionArgs" "DataFunctionArgs" --go
```

...and then removing duplicate imports of `DataFunctionArgs`.

This patch also fixes some `Handle` and `UIMatch` type complications. It
seems like the new Remix version comes with some better `RouteMatch`
types. This patch works to try and take advantage of that.

This patch also replaces all instances of `AppData` with `unknown`.

Ref: https://github.com/remix-run/remix/releases/tag/remix%402.0.0
Ref: https://remix.run/docs/en/main/start/v2