feat: save all prices and set `available` field
Instead of simply not saving unavailable prices (i.e. prices that have
sold out), we save them in database but set the `available` boolean flag
to false. This can be combined with the `createdAt` and `updatedAt`
field to show when a product last sold out at a specific retailer or
brand. This can be used to determine how popular a product is (i.e. if
the `available` field was set to `false` within five minutes of
`createdAt`, then we know the product sold out within five minutes).