fix(prisma): save specific product variants to sets
This patch updates the data model to save specific product variants to
sets instead of simply saving the entire product. This addresses an
oversight in a patchable way instead of trying to amend that previous
commit. I did not want to have to handle manual Prisma migrations, so
that is why this is a patch forward when ideally it should've been
amended unto the original commit.

Fixes: 240e3dcfda83 ("feat(prisma): products can be added to sets too")