feat: allow users to save products to sets
This patch reuses a bunch of code from the show view in the new product
page. Users can now click on a product image to save the specific
product variant displayed in that image to a set of their choice.

This feature will allow users to save products to the "own" and "want"
sets or any other wishlist name of their choice.