fix(scripts/save/shopify): save variant tags to db
This patch updates the Shopify script to save variant product tags to
Postgres. These are arbitrary tags assigned by the brand in Shopify.
This will be used to correlate products with specific collections or
styles.

This patch also adds some updates to ensure that the `upsert` behavior
is correct and we are updating or creating deeply nested relations (e.g.
the tags field on the product variants). Note that this is still not
perfect and there are some notable missing features: The variant price
field will not be updated if there is an additional price from a
different vendor at a different value. Instead, we skip that for now.