feat(prisma/schema): add posts and items data model
This patch adds a new `Post` model that represents a TikTok or Instagram
post imported by a user. Users can then link the post with product(s),
look(s), or item(s). This patch introduces a new `Item` model that
represents a product category (e.g. "white turtleneck sweater") that can
be used to generically associate products with each other and with
posts. Often, the exact brand/size of an item in a post is unknown. This
lets users annotate what is in a post without having to erroneously
associate it with an exact product. Only if the exact product is known
will it be associated with the post.

This allows users to upload fashion inspiration posts, label what types
of items are in each outfit, and then purchase products that satisfy the
item specifications.