feat: add jupyter notebooks for experimentation
This patch adds some Jupyter notebooks for quick experimentation (e.g.
to label the items in my saved Instagram posts).

I added the Python Prisma client to interact with my database in a
relatively type-safe manner in these notebooks.

This patch also updates the GitHub workflows to only generate the JS
Prisma client (as Python dependencies are not installed in CI).