feat(prisma): rename "set" model to "board"
This patch fixes an issue where the `Set` model name would conflict with
the built-in class names in Python and cause the Python Prisma client to
fail. This is a temporary fix, but I also think this will just be easier
to work with in general (as most languages have a `Set` class).

Ref: https://github.com/RobertCraigie/prisma-client-py/issues/842