fix(app/profile): own/want nav items go to sets
This patch updates the user profile page so that the "own" and "want"
navigation items simply open the corresponding set page. There is
nothing particularly special about these sets, so I figured I might as
well have them included just as normal sets and exposed so to the end
user.

This does result in some unintuitive behavior where both the "sets" and
the "own/want" tab will be active at a given time. But it makes sense.