fix(app): replace `Set` data model with `Board`
This patch updates the underlying app implementation to use the newly
renamed `Board` data model instead of the `Set` data model. Note that
this does not update any user-facing naming: Users will still see "sets"
in the UI. Variables are also still named based on sets.

This was the minimal change required to get the app working with the
latest rename Prisma migration.

Fixes: 2f74c5cd1c3f ("feat(prisma): rename "set" model to "board"")