fix(app/filters): add new relation filter types
This patch updates the filter types to include the new item and post
types. Previously, those new types were causing TS inference to fail.

This patch also hides those new filter relations from the UI (as they're
not useful to filter on at the moment and there are no routes to load
the filter options).

Fixes: 42c9c6c509a2 ("feat(prisma/schema): add posts and items data model")