feat(app): add wardrobe panel open/close button
This patch adds a button to toggle the wardrobe side-panel open and
closed. This side-panel is not actually that useful in practice (as it
is implemented right now), but I do not want to delete the code paths,
as I do want a side-panel where users can save and curate what they're
looking at to their profile (i.e. "their wardrobe").