fix(remix-ui): install correct `pnpm` version
This patch ensures that we still take advantage of the layer caching
that only relies on the `pnpm-lock.yaml` file and not the `package.json`
file (which could change due to added scripts, etc). Previously, there
was an issue with the default `pnpm` version not being compatible with
the `pnpm` version specified by our `package.json`. I now explicitly
install the same version with `corepack` before the `pnpm fetch` step.

Alternatively, I should probably just update the `pnpm` version used by
the project to match what is used in the Node.js 22 Docker image by
default.

Ref: https://pnpm.io/cli/fetch

Fixes: 49e0ac2b1b94 ("fix: reckless change for docker build")