feat(svelte-ui): add basic collections list page
This patch adds a very basic collections list page that connects to our
Postgres database and lists all of the collections available. The data
is streamed from the server [[1]] via a server-only load function [[2]].

I use Drizzle [[3]] as the ORM of choice. It is more performant than
Prisma and more flexible [[4]]. I introspect an existing Postgres
database that contains all of the data and schemas currently in
production (this database state was seeded with Prisma from the code in
`remix-ui`) [[5]].

[1]: https://svelte.dev/docs/kit/load#Streaming-with-promises
[2]: https://svelte.dev/docs/kit/load#Page-data
[3]: https://orm.drizzle.team/
[4]: https://orm.drizzle.team/docs/overview
[5]: https://orm.drizzle.team/docs/get-started/postgresql-existing#step-4---introspect-your-database